python/pyproject.toml
[project]
name = "c0lor-mem-backend"
version = "0.1.0"
description = "Python backend for c0lor-mem display test image generator"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pillow>=11.0.0",
    "pillow-heif>=1.2.0",
    "colour-science>=0.4.6",
    "pydantic>=2.10.0",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
    "pyinstaller>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
